refactor: make it work with latest Bazel rules

+++ typescript/test/test-application/test-ts.ts
@@ -12,13 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import * as util from 'util';
 import * as child_process from 'child_process';
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import {describe, it} from 'mocha';
 
-const exec = util.promisify(child_process.exec);
 const root = process.cwd();
 const baselineZip = path.join(
   root,
@@ -32,39 +30,67 @@ const packedLib = 'showcase-0.1.0.tgz';
 const packedLibPath = path.join(showcaseLib, packedLib);
 const testFixtures = path.join(root, 'test-fixtures');
 const protos = path.join(testFixtures, 'protos');
-const jsTestApplication = path.join(testFixtures, 'test-application-js');
-const localJsApplication = path.join(root, 'test-application-js');
+const tsTestApplication = path.join(testFixtures, 'test-application-ts');
+const localTsApplication = path.join(root, 'test-application-ts');
 
-describe('Test application for JavaScript users', () => {
+function spawn(cmd: string, args: string[]) {
+  return new Promise<void>((resolve, reject) => {
+    const child = child_process.spawn(cmd, args);
+    child.stdout.pipe(process.stdout);
+    child.stderr.pipe(process.stderr);
+    child.stdin.end();
+    child.on('close', (code, signal) => {
+      if (code !== 0 || signal) {
+        reject(
+          new Error(`Process ${cmd} failed: code ${code}, signal ${signal}`)
+        );
+      } else {
+        resolve();
+      }
+    });
+  });
+}
+
+describe('Test application for TypeScript users', () => {
   it('unzip showcase test output', async function () {
     this.timeout(240000);
     process.chdir(root);
-    await exec(`unzip -o "${baselineZip}" '.test-out-showcase/*' -d .`);
+    await spawn('unzip', [
+      '-o',
+      baselineZip,
+      '.test-out-showcase/*',
+      '-d',
+      '.',
+    ]);
   });
   it('npm install showcase', async function () {
     this.timeout(240000);
     // copy protos to generated client library and copy test application to local.
-    fs.copySync(protos, path.join(showcaseLib, 'protos'));
-    fs.copySync(jsTestApplication, localJsApplication);
+    if (!fs.existsSync(path.join(showcaseLib, 'protos'))) {
+      fs.copySync(protos, path.join(showcaseLib, 'protos'));
+    }
+    if (!fs.existsSync(localTsApplication)) {
+      fs.copySync(tsTestApplication, localTsApplication);
+    }
     process.chdir(showcaseLib);
-    await exec('npm install');
+    await spawn('npm', ['install']);
   });
   it('npm pack showcase library and copy it to test application', async function () {
     this.timeout(240000);
-    await exec('npm pack');
-    fs.copySync(packedLibPath, path.join(localJsApplication, packedLib));
+    await spawn('npm', ['pack']);
+    fs.copySync(packedLibPath, path.join(localTsApplication, packedLib));
   });
   it('npm install showcase library in test application', async function () {
     this.timeout(240000);
-    process.chdir(localJsApplication);
-    await exec('npm install --legacy-peer-deps');
+    process.chdir(localTsApplication);
+    await spawn('npm', ['install', '--legacy-peer-deps']);
   });
   it('run integration in test application', async function () {
     this.timeout(240000);
-    await exec('npm test');
+    await spawn('npm', ['test']);
   });
   it('run browser test in application', async function () {
     this.timeout(240000);
-    await exec('npm run browser-test');
+    await spawn('npm', ['run', 'browser-test']);
   });
 });
